authorization-admin

+++ app/Http/Controllers/Api/Admin/OrderController.php
@@ -13,6 +13,19 @@ use Symfony\Component\HttpFoundation\Response;
 
 class OrderController extends ApiController
 {
+
+    protected $account;
+
+    /**
+     * CategoryController constructor..
+     *
+     * @param UploadImageService   $uploadImageService   UploadImageService
+     */
+    public function __construct()
+    {
+        $this->account = auth('api')->user();
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -21,10 +34,13 @@ class OrderController extends ApiController
     public function index(Request $request)
     {
         try {
-            $order = Order::with(['user','payment','coupon'])->orderFilter($request)->orderBy('created_at', 'desc')->paginate(config('paginate.number_orders'));
-            return $this->formatPaginate($order);
+            $orders = Order::with(['user','payment','coupon'])->orderFilter($request)->orderBy('created_at', 'desc')->paginate(config('paginate.number_orders'));
+            if ($this->account->can('view', Order::all()->first())) {
+                return $this->formatPaginate($orders);
+            } else {
+                return $this->errorResponse(config('define.no_authorization'), Response::HTTP_UNAUTHORIZED);
+            }
         } catch (Exception $ex) {
-            dd($ex->getMessage());
             return $this->errorResponse("Orders can not be show.", Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
@@ -40,11 +56,14 @@ class OrderController extends ApiController
     {
         try {
             $order = Order::with(['user','payment','coupon','orderDetails.product:id,name,price'])->findOrFail($id);
-            return $this->successResponse($order, Response::HTTP_OK);
+            if ($this->account->can('view', $order)) {
+                return $this->successResponse($order, Response::HTTP_OK);
+            } else {
+                return $this->errorResponse(config('define.no_authorization'), Response::HTTP_UNAUTHORIZED);
+            }
         } catch (ModelNotFoundException $ex) {
             return $this->errorResponse("Order not found.", Response::HTTP_NOT_FOUND);
         } catch (Exception $ex) {
-            dd($ex->getMessage());
             return $this->errorResponse("Occour error when show Order.", Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
@@ -60,14 +79,18 @@ class OrderController extends ApiController
     public function update(UpdateOrderRequest $request, Order $order)
     {
         try {
-            if ($order->processing_status == Order::STATUS_PROCESSING) {
-                $order->processing_status = $request->processing_status;
+            if ($this->account->can('update', $order)) {
+                if ($order->processing_status == Order::STATUS_PROCESSING) {
+                    $order->processing_status = $request->processing_status;
+                }
+                $order->payment_status = $request->payment_status;
+                $order->delivery_time = $request->delivery_time;
+                $order->save();
+    
+                return $this->successResponse("Update order successfully", Response::HTTP_OK);
+            } else {
+                return $this->errorResponse(config('define.no_authorization'), Response::HTTP_UNAUTHORIZED);
             }
-            $order->payment_status = $request->payment_status;
-            $order->delivery_time = $request->delivery_time;
-            $order->save();
-
-            return $this->successResponse("Update order successfully", Response::HTTP_OK);
         } catch (Exception $ex) {
             return $this->errorResponse("Occour error when show Order.", Response::HTTP_INTERNAL_SERVER_ERROR);
         }
@@ -85,13 +108,17 @@ class OrderController extends ApiController
     public function destroy(Order $order)
     {
         try {
-            $order->orderDetails()->delete();
-            $order->delete();
-            return $this->successResponse("Delete order successfully", Response::HTTP_OK);
+            if ($this->account->can('delete', $order)) {
+                $order->orderDetails()->delete();
+                $order->delete();
+                return $this->successResponse("Delete order successfully", Response::HTTP_OK);
+            } else {
+                return $this->errorResponse(config('define.no_authorization'), Response::HTTP_UNAUTHORIZED);
+            }
+            
         } catch (ModelNotFoundException $ex) {
             return $this->errorResponse("Order not found.", Response::HTTP_NOT_FOUND);
         } catch (Exception $ex) {
-            dd($ex->getMessage());
             return $this->errorResponse("Occour error when show order.", Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
